added guards, valdation pipes
