+++ src/applicants/applicant.dto.ts
@@ -1,12 +1,41 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsEmail, IsOptional, IsString } from 'class-validator';
+
 export class CreateApplicantDto {
+  @IsString()
   username: string;
+
+  @IsString()
   firstName: string;
+
+  @IsString()
   lastName: string;
+
+  @IsEmail()
   email: string;
+
+  @IsString()
   password: string;
 }
 
 export class LoginApplicantDTO {
   username: string;
   password: string;
+}
+
+export class UpdateApplicantDto {
+  @ApiProperty({ required: false })
+  @IsOptional()
+  @IsString()
+  firstName?: string;
+
+  @ApiProperty({ required: false })
+  @IsOptional()
+  @IsString()
+  lastName?: string;
+
+  @ApiProperty({ required: false })
+  @IsOptional()
+  @IsEmail()
+  email?: string;
 }
