+++ src/applicants/applicants.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, Logger } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  Logger,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Applicant } from './applicant.entity';
@@ -23,14 +28,35 @@ export class ApplicantsService {
     applicant = await this.applicantsRepository.save(applicant);
     return applicant;
   }
-  async updateOne(username: string, updateData: Partial<Applicant>): Promise<Applicant> {
+
+  async updateOne(
+    username: string,
+    updateData: Partial<Applicant>,
+  ): Promise<Applicant> {
     const applicant = await this.applicantsRepository.findOne({
       where: { username },
     });
+
     if (!applicant) {
-      throw new Error(`Applicant with username ${username} not found`);
+      throw new NotFoundException(
+        `Applicant with username ${username} not found`,
+      );
     }
-    Object.assign(applicant, updateData);
+
+    // Add validation logic here
+    if (updateData.hasOwnProperty('username')) {
+      throw new BadRequestException('Updating username is not allowed');
+    }
+
+    // Only allow updates to these fields
+    const allowedFields = ['firstName', 'lastName', 'email'];
+
+    const filteredUpdateData = Object.fromEntries(
+      Object.entries(updateData).filter(([key]) => allowedFields.includes(key)),
+    );
+
+    Object.assign(applicant, filteredUpdateData);
+
     return await this.applicantsRepository.save(applicant);
   }
 }
